refactor(metadata): migrate inventoryagent component to v2
