+++ cmd/systray/command/command.go
@@ -28,7 +28,7 @@ import (
 	nooptagger "github.com/DataDog/datadog-agent/comp/core/tagger/fx-noop"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	haagentfx "github.com/DataDog/datadog-agent/comp/haagent/fx"
-	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/inventoryagentimpl"
+	inventoryagentfx "github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/fx"
 	logscompressionfx "github.com/DataDog/datadog-agent/comp/serializer/logscompression/fx"
 	systray "github.com/DataDog/datadog-agent/comp/systray/systray/def"
 	systrayfx "github.com/DataDog/datadog-agent/comp/systray/systray/fx"
@@ -113,7 +113,7 @@ func MakeCommand() *cobra.Command {
 				logscompressionfx.Module(),
 				nooptagger.Module(),
 				// We need inventoryagent to fill the status page generated by the flare.
-				inventoryagentimpl.Module(),
+				inventoryagentfx.Module(),
 				// inventoryagent require a serializer. Since we're not actually sending the payload to
 				// the backend a nil will work.
 				fx.Provide(func() serializer.MetricSerializer {
