+++ comp/metadata/inventoryagent/def/component.go
