+++ comp/metadata/inventoryagent/impl/status.go
