+++ comp/metadata/inventoryagent/impl/status_templates/inventory.tmpl
