+++ comp/metadata/inventoryagent/impl/status_templates/inventoryHTML.tmpl
